separate lyrics hook
