+++ client/hooks/useLyrics.ts
@@ -1,8 +1,24 @@
 import { useAuth0 } from '@auth0/auth0-react'
-import { useQueryClient, useMutation } from '@tanstack/react-query'
+import { useQueryClient, useMutation, useQuery } from '@tanstack/react-query'
 import { useParams } from 'react-router-dom'
 import { LyricsData } from '../../models/lyrics'
-import { addLyrics } from '../apis/lyrics'
+import { addLyrics, editLyrics, getLyrics } from '../apis/lyrics'
+import { EditedLyrics } from '../../models/songs'
+
+export function useLyrics(songId: number, collectionId: number) {
+  return useQuery({
+    queryKey: ['lyrics', songId, collectionId],
+    queryFn: () => getLyrics(songId, collectionId),
+  })
+}
+
+export function useEditLyrics() {
+  const qc = useQueryClient()
+  return useMutation({
+    mutationFn: (lyrics: EditedLyrics) => editLyrics(lyrics),
+    onSuccess: () => qc.invalidateQueries({ queryKey: ['lyrics'] }),
+  })
+}
 
 export function useAddLyrics() {
   const qc = useQueryClient()
